+++ anybox/recipe/openerp/base.py
@@ -451,11 +451,18 @@ class BaseRecipe(object):
     def develop(self, src_directory, setup_has_pil=False):
         """Develop the specified source distribution.
 
-        Any call to zc.recipe.eggs will use that developped version.
-        develop() launches a subprocess, to which we need to forward
+        Any call to ``zc.recipe.eggs`` will use that developped version.
+        :meth:`develop` launches a subprocess, to which we need to forward
         the paths to requirements via PYTHONPATH.
-        If setup_has_pil is True, an altered version of setup that does not
-        require it is produced to perform the develop.
+
+        :param setup_has_pil: if ``True``, an altered version of setup that
+                              does not require PIL is produced to perform the
+                              develop, so that installation can be done with
+                              ``Pillow`` instead. Recent enough versions of
+                              OpenERP/Odoo are directly based on Pillow.
+        :returns: project name of the distribution that's been "developed"
+                  This is useful for OpenERP/Odoo itself, whose project name
+                  changed within the 8.0 stable branch.
         """
         logger.debug("Developing %r", src_directory)
         develop_dir = self.b_options['develop-eggs-directory']
@@ -468,15 +475,24 @@ class BaseRecipe(object):
             setup = src_directory
 
         try:
-            zc.buildout.easy_install.develop(setup, develop_dir)
+            egg_link = zc.buildout.easy_install.develop(setup, develop_dir)
         finally:
             if setup_has_pil:
                 os.unlink(setup)
+
+        suffix = '.egg-link'
 
         if pythonpath_bak is None:
             os.unsetenv('PYTHONPATH')
         else:
             os.putenv('PYTHONPATH', pythonpath_bak)
+
+        if not egg_link.endswith(suffix):
+            raise RuntimeError(
+                "Development of OpenERP/Odoo distribution "
+                "produced an unexpected egg link: %r" % egg_link)
+
+        return os.path.basename(egg_link)[:-len(suffix)]
 
     def parse_addons(self, options):
         """Parse the addons options into :attr:`sources`.
@@ -1258,11 +1274,16 @@ class BaseRecipe(object):
 
         These can be, e.g., forbidden by the freeze process."""
 
-        # GR TODO this will break with OSError if main package is renamed to
-        # 'odoo' we'll see then what the needed correction exactly is rather
-        # than swallowing the exception now
-        shutil.rmtree(join(self.openerp_dir, 'openerp.egg-info'))
-        # setup rewritten without PIL is cleaned during the process itself
+        # from here we can't guess whether it's 'openerp' or 'odoo'.
+        # Nothing guarantees that this method is called after develop().
+        # It is in practice now, but one day, the extraction as a separate
+        # script of freeze/extract will become a reality.
+        for proj_name in ('openerp', 'odoo'):
+            egg_info_dir = join(self.openerp_dir, proj_name + '.egg-info')
+            if os.path.exists(egg_info_dir):
+                shutil.rmtree(egg_info_dir)
+        # if setup.py has been rewritten without PIL, it is cleaned
+        # during the process itself
 
     def buildout_cfg_name(self, argv=None):
         """Return the name of the config file that's been called.